requirements.txt
aiohttp
aiodns
aiofiles
backoff
Django~=2.2.14  # LTS version, switch only if we have a compelling reason to
django-filter~=2.3.0
django-import-export~=2.2.0
djangorestframework~=3.10.2
djangorestframework-queryfields~=1.0.0
drf-nested-routers~=0.91.0
drf-spectacular>=0.9.11
dynaconf>=2.2,<4.0
gunicorn>=19.9,<20.1
pygtrie~=2.3.3
psycopg2>=2.7,<2.9
PyYAML>=5.1.1,<5.4.0
python-gnupg~=0.4.6
redis>=3.4.0
rq>=1.1,<1.5
setuptools>=39.2.0
whitenoise>=4.1.3,<5.2.0
